Sayantan was getting weird results; turns out it's because of a weird bug in imfilter() for small matrices that are singles.  Switched all the functions back to using doubles, (I missed some previously) and forced VELfilt to cast the inputs to doubles before using imfilter, then cast back to the original type once the danger is past.

+++ IMgrid.m
@@ -1,7 +1,7 @@
 function [X,Y]=IMgrid(L,S,G,imClass)
 % --- Grid Generation Subfunction ---
 if nargin<4
-    imClass = 'single';
+    imClass = 'double';
 end
 
 %grid buffer
